ISSUE-N/A Introduce Sphere/Hitable/HitableList
Model Sphere as a class and remove the hardcoded Sphere from
the color method. We now raytrace 2 spheres as part of the demos
using the normal vector at the intersection point to derive the
color of the point.

I've introduced the use of std::vector and std::uniq_ptr instead
of using raw pointers and arrays as the original book does.